requirements.txt
python-dotenv==1.1.1
langchain==0.3.27
langchain-huggingface==0.3.1
langgraph==0.6.7
openai==1.107.1